# swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "info": {
    "title": "Nightscout API",
    "description": "Own your DData with the Nightscout API",
    "version": "14.0.6",
    "license": {
      "name": "AGPL 3",
      "url": "https://www.gnu.org/licenses/agpl.txt"
    }
  },
  "security": [
    {
      "api_secret": [],
      "token_in_url": [],
      "jwtoken": []
    }
  ],
  "paths": {
...
    "/treatments": {
      "get": {
        "summary": "Treatments",
        "description": "The Treatments endpoint returns information about the Nightscout treatments.",
        "tags": [
          "Treatments"
        ],
        "parameters": [
          {
            "name": "find",
            "in": "query",
            "description": "The query used to find entries, supports nested query syntax.  Examples `find[insulin][$gte]=3` `find[carb][$gte]=100` `find[eventType]=Correction+Bolus` All find parameters are interpreted as strings.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Number of entries to return.",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of treatments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Treatments"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Treatments"
        ],
        "summary": "Add new treatments.",
        "description": "",
        "operationId": "addTreatments",
        "responses": {
          "200": {
            "description": "Rejected list of treatments.  Empty list is success."
          },
          "405": {
            "description": "Invalid input"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Treatments"
              }
            }
          },
          "description": "Treatments to be uploaded.",
          "required": true
        }
      },
      "delete": {
        "tags": [
          "Treatments"
        ],
        "summary": "Delete treatments matching query.",
        "description": "Remove treatments, same search syntax as GET.",
        "operationId": "remove",
        "parameters": [
          {
            "name": "find",
            "in": "query",
            "description": "The query used to find treatments to delete, support nested query syntax, for example `find[insulin][$gte]=3`. All find parameters are interpreted as strings.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Number of entries to return.",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Empty list is success."
          }
        }
      }
    },
    "/treatments/{spec}": {
      "delete": {
        "summary": "Delete treatments record with id provided in spec",
        "description": "The Treatments endpoint returns information about the\nNightscout devicestatus records.\n",
        "parameters": [
          {
            "name": "spec",
            "in": "path",
            "description": "treatment id, such as `55cf81bc436037528ec75fa5`\n",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Treatments"
        ],
        "responses": {
          "200": {
            "description": "A status record of the delete.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteStatus"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/devicestatus/": {
      "get": {
        "summary": "All Devicestatuses matching query",
        "description": "The Devicestatus endpoint returns information about the Nightscout devicestatus records.",
        "parameters": [
          {
            "name": "find",
            "in": "query",
            "description": "The query used to find entries, support nested query syntax, for example `find[dateString][$gte]=2015-08-27`.  All find parameters are interpreted as strings.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Number of devicestatus records to return.",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "tags": [
          "Devicestatus"
        ],
        "responses": {
          "200": {
            "description": "An array of devicestatus entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Devicestatuses"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Devicestatus"
        ],
        "summary": "Add new devicestatus records.",
        "description": "",
        "operationId": "addDevicestatuses",
        "responses": {
          "200": {
            "description": "Rejected list of device statuses.  Empty list is success."
          },
          "405": {
            "description": "Invalid input"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Devicestatuses"
              }
            }
          },
          "description": "Device statuses to be uploaded.",
          "required": true
        }
      },
      "delete": {
        "summary": "Delete all Devicestatus records matching query",
        "description": "The Devicestatus endpoint returns information about the\nNightscout devicestatus records.\n",
        "parameters": [
          {
            "name": "find",
            "in": "query",
            "description": "The query used to find entries, support nested query syntax, for\nexample `find[created_at][$gte]=2015-08-27`.  All find parameters\nare interpreted as strings.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Devicestatus"
        ],
        "responses": {
          "200": {
            "description": "A status record of the delete.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteStatus"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/devicestatus/{spec}": {
      "delete": {
        "summary": "Delete devicestatus record with id provided in spec",
        "description": "The Devicestatus endpoint returns information about the\nNightscout devicestatus records.\n",
        "parameters": [
          {
            "name": "spec",
            "in": "path",
            "description": "entry id, such as `55cf81bc436037528ec75fa5`\n",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Devicestatus"
        ],
        "responses": {
          "200": {
            "description": "A status record of the delete.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteStatus"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "api_secret": {
        "type": "apiKey",
        "name": "api_secret",
        "in": "header",
        "description": "The hash of the API_SECRET env var"
      },
      "token_in_url": {
        "type": "apiKey",
        "name": "token",
        "in": "query",
        "description": "Add token as query item in the URL. You can manage access Token in `/admin`. This uses json webtokens."
      },
      "jwtoken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use this if you know the temporary json webtoken.",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Devicestatus": {
        "required": [
          "device",
          "created_at"
        ],
        "properties": {
          "device": {
            "type": "string",
            "description": "Device type and hostname for example openaps://hostname"
          },
          "created_at": {
            "type": "string",
            "description": "dateString, prefer ISO `8601`"
          },
          "openaps": {
            "type": "string",
            "description": "OpenAPS devicestatus record - TODO: Fill Out Details"
          },
          "loop": {
            "type": "string",
            "description": "Loop devicestatus record - TODO: Fill Out Details"
          },
          "pump": {
            "$ref": "#/components/schemas/pump"
          },
          "uploader": {
            "$ref": "#/components/schemas/uploader"
          },
          "xdripjs": {
            "$ref": "#/components/schemas/xdripjs"
          }
        }
      },
      "Devicestatuses": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Devicestatus"
        }
      },
      "pump": {
        "properties": {
          "clock": {
            "type": "string",
            "description": "dateString, prefer ISO `8601`"
          },
          "battery": {
            "$ref": "#/components/schemas/pumpbattery"
          },
          "reservoir": {
            "type": "number",
            "description": "Amount of insulin remaining in pump reservoir"
          },
          "status": {
            "$ref": "#/components/schemas/pumpstatus"
          }
        }
      },
      "pumpbattery": {
        "properties": {
          "status": {
            "type": "string",
            "description": "Pump Battery Status String"
          },
          "voltage": {
            "type": "number",
            "description": "Pump Battery Voltage Level"
          }
        }
      },
      "pumpstatus": {
        "properties": {
          "status": {
            "type": "string",
            "description": "Pump Status String"
          },
          "bolusing": {
            "type": "boolean",
            "description": "Is Pump Bolusing"
          },
          "suspended": {
            "type": "boolean",
            "description": "Is Pump Suspended"
          },
          "timestamp": {
            "type": "string",
            "description": "dateString, prefer ISO `8601`"
          }
        }
      },
      "uploader": {
        "properties": {
          "batteryVoltage": {
            "type": "number",
            "description": "Uploader Device Battery Voltage"
          },
          "battery": {
            "type": "number",
            "description": "Uploader Device Battery Percentage Charge Remaining"
          }
        }
      },
      "xdripjs": {
        "properties": {
          "state": {
            "type": "number",
            "description": "CGM Sensor Session State Code"
          },
          "stateString": {
            "type": "string",
            "description": "CGM Sensor Session State String"
          },
          "stateStringShort": {
            "type": "string",
            "description": "CGM Sensor Session State Short String"
          },
          "txId": {
            "type": "string",
            "description": "CGM Transmitter ID"
          },
          "txStatus": {
            "type": "number",
            "description": "CGM Transmitter Status"
          },
          "txStatusString": {
            "type": "string",
            "description": "CGM Transmitter Status String"
          },
          "txStatusStringShort": {
            "type": "string",
            "description": "CGM Transmitter Status Short String"
          },
          "txActivation": {
            "type": "number",
            "description": "CGM Transmitter Activation Milliseconds After Epoch"
          },
          "mode": {
            "type": "string",
            "description": "Mode xdrip-js Application Operationg: expired, not expired, etc."
          },
          "timestamp": {
            "type": "number",
            "description": "Last Update Milliseconds After Epoch"
          },
          "rssi": {
            "type": "number",
            "description": "Receive Signal Strength of Transmitter"
          },
          "unfiltered": {
            "type": "number",
            "description": "Most Recent Raw Unfiltered Glucose"
          },
          "filtered": {
            "type": "number",
            "description": "Most Recent Raw Filtered Glucose"
          },
          "noise": {
            "type": "number",
            "description": "Calculated Noise Value - 1=Clean, 2=Light, 3=Medium, 4=Heavy"
          },
          "noiseString": {
            "type": "number",
            "description": "Noise Value String"
          },
          "slope": {
            "type": "number",
            "description": "Calibration Slope Value"
          },
          "intercept": {
            "type": "number",
            "description": "Calibration Intercept Value"
          },
          "calType": {
            "type": "string",
            "description": "Algorithm Used to Calculate Calibration Values"
          },
          "lastCalibrationDate": {
            "type": "number",
            "description": "Most Recent Calibration Milliseconds After Epoch"
          },
          "sessionStart": {
            "type": "number",
            "description": "Sensor Session Start Milliseconds After Epoch"
          },
          "batteryTimestamp": {
            "type": "number",
            "description": "Most Recent Batter Status Read Milliseconds After Epoch"
          },
          "voltagea": {
            "type": "number",
            "description": "Voltage of Battery A"
          },
          "voltageb": {
            "type": "number",
            "description": "Voltage of Battery B"
          },
          "temperature": {
            "type": "number",
            "description": "Transmitter Temperature"
          },
          "resistance": {
            "type": "number",
            "description": "Sensor Resistance"
          }
        }
      },
      "Treatment": {
        "properties": {
          "_id": {
            "type": "string",
            "description": "Internally assigned id."
          },
          "eventType": {
            "type": "string",
            "description": "The type of treatment event."
          },
          "created_at": {
            "type": "string",
            "description": "The date of the event, might be set retroactively ."
          },
          "glucose": {
            "type": "string",
            "description": "Current glucose."
          },
          "glucoseType": {
            "type": "string",
            "description": "Method used to obtain glucose, Finger or Sensor."
          },
          "carbs": {
            "type": "number",
            "description": " Amount of carbs consumed in grams."
          },
          "protein": {
            "type": "number",
            "description": " Amount of protein consumed in grams."
          },
          "fat": {
            "type": "number",
            "description": " Amount of fat consumed in grams."
          },
          "insulin": {
            "type": "number",
            "description": "Amount of insulin, if any."
          },
          "units": {
            "type": "string",
            "description": "The units for the glucose value, mg/dl or mmol."
          },
          "notes": {
            "type": "string",
            "description": "Description/notes of treatment."
          },
          "enteredBy": {
            "type": "string",
            "description": "Who entered the treatment."
          }
        }
      },
      "Treatments": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Treatment"
        }
      },
...
      "Error": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "object"
          }
        }
      },
      "DeleteStatus": {
        "properties": {
          "n": {
            "type": "integer",
            "format": "int32",
            "description": "Number of records deleted"
          },
          "optime": {
            "$ref": "#/components/schemas/optime"
          },
          "electionId": {
            "type": "string",
            "description": "Election id of operation"
          },
          "ok": {
            "type": "integer",
            "format": "int32",
            "description": "Status of whether delete was successful"
          },
          "operationTime": {
            "type": "string",
            "description": "Time delete operation was executed"
          },
          "$clusterTime": {
            "type": "string",
            "description": "Information about execution time in cluster environment"
          }
        }
      },
      "optime": {
        "properties": {
          "ts": {
            "type": "string",
            "description": "Time the operation started"
          },
          "t": {
            "type": "integer",
            "format": "int32",
            "description": "Time the operation took to complete"
          }
        }
      }
    }
  }
}